pyproject.toml
[project]
name = "agent-docs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.8.0",
    "python-docx>=1.1.2",
]
